requirements.txt
openai
phidata
python-dotenv
duckduckgo-search
yfinance
packaging
fastapi
uvicorn
groq
python-multipart